Improve func_8012F2DC_13E28C to CURRENT(2200)

+++ src.us/overlay_gameplay/outside/13DA70.c
@@ -261,7 +261,7 @@ void func_8012F24C_13E1FC(u8 *arg0, s32 arg1)
 #pragma GLOBAL_ASM("asm/nonmatchings/overlay_gameplay/outside/13DA70/func_8012F24C_13E1FC.s")
 #endif
 
-// CURRENT(2415)
+// CURRENT(2200)
 #ifdef NON_MATCHING
 void func_8012F2DC_13E28C(s32 arg0, s32 arg1, s32 arg2)
 {
@@ -269,7 +269,7 @@ void func_8012F2DC_13E28C(s32 arg0, s32 arg1, s32 arg2)
 	f32 var_f0;
 	s32 var_v1;
 	u8 temp_v0;
-	u16 temp_a0;
+	u32 temp_a0;
 
 	temp_a0 = (D_80140C78_14FC28 * 0x3E8) & 0xFFFF;
 	if (((f64) (f32) coss(temp_a0) / 32768.0) >= 0.0) {
